Fix build
Revert SystemUI: Add global reboot option
Revert QuickBoot: switch normal power off to quickboot
Fixed conflicts with MSim commits

+++ packages/SystemUI/src/com/android/systemui/statusbar/policy/MSimNetworkControllerImpl.java
@@ -434,7 +434,7 @@ public class MSimNetworkControllerImpl extends NetworkControllerImpl {
                 }
         } else if (action.equals(TelephonyIntents.ACTION_DEFAULT_DATA_SUBSCRIPTION_CHANGED)) {
             // Update data in QS
-            long subId = intent.getLongExtra(PhoneConstants.SUBSCRIPTION_KEY, -1);
+            int subId = intent.getIntExtra(PhoneConstants.SUBSCRIPTION_KEY, -1);
 
             if (subId == -1) {
                 Slog.e(TAG, "No subId in ACTION_DEFAULT_DATA_SUBSCRIPTION_CHANGED");
